mando user_id en los endpoints necesarios

+++ src/services/conferenceServices.ts
@@ -72,12 +72,23 @@ const handleConferenceError = (err: any, isCreate: boolean) => {
 
 export const createConference = async (
   conferencia: Omit<Conference, 'id'>,
-  chairs: User[]
+  chairs: User[],
+  user_id: string
 ): Promise<Conference> => {
   try {
+
+    //prueba para agregar admin
+    // const response = await api.post('/user/register-admin/', {
+    //   full_name: "nicolas admin",
+    //   affiliation: "unlp",
+    //   email: "nicoarditi13@gmail.com",
+    //   password: "48595519"
+    // });
+
     const response = await api.post('/api/conference/', {
       ...conferencia,
       chairs: chairs.map(user => user.id),
+      user_id
     });
 
     return response.data;
@@ -90,12 +101,14 @@ export const createConference = async (
 export const updateConference = async (
   id: string,
   conferencia: Omit<Conference, 'id'>,
-  chairs: User[]
+  chairs: User[],
+  user_id: string
 ): Promise<Conference> => {
   try {
     const response = await api.patch(`/api/conference/${id}/`, {
       ...conferencia,
       chairs: chairs.map(user => user.id),
+      user_id
     });
     return response.data;
   } catch (err) {
@@ -104,6 +117,6 @@ export const updateConference = async (
   }
 };
 
-export const deleteConference = async (id: string): Promise<void> => {
-  await api.delete(`/api/conference/${id}/`);
+export const deleteConference = async (id: string, user_id: string): Promise<void> => {
+  await api.delete(`/api/conference/${id}/?user_id=${user_id}`);
 };
